requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
dataclasses-json==0.6.1
exceptiongroup==1.1.3
frozenlist==1.4.0
google-api-core==2.13.0
google-auth==2.23.4
google-cloud-core==2.3.3
google-cloud-speech==2.22.0
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.2
grpcio-status==1.59.2
idna==3.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.331
langsmith==0.0.60
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.1
openai==0.28.1
packaging==23.2
proto-plus==1.22.3
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
sniffio==1.3.0
SQLAlchemy==2.0.23
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.7
yarl==1.9.2
